+++ src/components/view/main/decision/decision-view.ts
