+++ src/components/view/main/decision/decision.scss
